+++ crawl-ref/source/state.h
@@ -47,7 +47,7 @@ struct game_state
                              // take action to handle it.
     time_t last_winch;       // Time of last resize, for crash dumps.
 
-    uint32_t seed;
+    uint64_t seed;
 
     bool io_inited;         // Is curses or the equivalent initialised?
     bool need_save;         // Set to true when game can be saved, false when the game ends.
